# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ecommerce Platform API",
    "version": "1.0.0",
    "description": "REST API for the ecommerce platform catalog, auth, cart, orders, health, and uploads."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from POST /api/auth/login or /api/auth/refresh"
      }
    },
...
  },
  "paths": {
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "operationId": "listProduct",
        "summary": "List products",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "categoryId",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "minimum": 0
            },
            "required": false,
            "name": "maxPrice",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "minimum": 0
            },
            "required": false,
            "name": "minPrice",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "price_asc",
                "price_desc",
                "title_asc",
                "title_desc"
              ]
            },
            "required": false,
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "sku": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "compareAtPrice": {
                            "type": "number"
                          },
                          "stock": {
                            "type": "integer"
                          },
                          "status": {
                            "type": "string",
                            "enum": ["draft", "published", "archived"]
                          },
                          "categoryId": {
                            "type": "string"
                          },
                          "categoryName": {
                            "type": "string"
                          },
                          "brandId": {
                            "type": "string"
                          },
                          "brandName": {
                            "type": "string"
                          },
                          "images": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "attributes": {
                            "type": "object",
                            "additionalProperties": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "array",
                                  "items": {
                                    "type": "string"
                                  }
                                }
                              ]
                            }
                          },
                          "metadata": {
                            "type": "object",
                            "additionalProperties": {
                              "nullable": true
                            }
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "slug",
                          "sku",
                          "description",
                          "price",
                          "stock",
                          "status",
                          "images",
                          "attributes",
                          "metadata",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  },
                  "required": ["data", "total", "page", "limit"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "operationId": "createProduct",
        "summary": "Create product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "sku": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 5000,
                    "default": ""
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0
                  },
                  "compareAtPrice": {
                    "type": "number",
                    "minimum": 0
                  },
                  "stock": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 0
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published", "archived"],
                    "default": "draft"
                  },
                  "categoryId": {
                    "type": "string"
                  },
                  "brandId": {
                    "type": "string"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "default": []
                  },
                  "attributes": {
                    "type": "object",
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    },
                    "default": {}
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    },
                    "default": {}
                  }
                },
                "required": ["name", "price"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "sku": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "compareAtPrice": {
                      "type": "number"
                    },
                    "stock": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["draft", "published", "archived"]
                    },
                    "categoryId": {
                      "type": "string"
                    },
                    "categoryName": {
                      "type": "string"
                    },
                    "brandId": {
                      "type": "string"
                    },
                    "brandName": {
                      "type": "string"
                    },
                    "images": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "attributes": {
                      "type": "object",
                      "additionalProperties": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        ]
                      }
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": {
                        "nullable": true
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "sku",
                    "description",
                    "price",
                    "stock",
                    "status",
                    "images",
                    "attributes",
                    "metadata",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": ["error", "details"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "operationId": "listCategory",
        "summary": "List categorys",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "image": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": ["draft", "published"]
                          },
                          "productCount": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "slug",
                          "image",
                          "status",
                          "productCount",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  },
                  "required": ["data", "total", "page", "limit"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "operationId": "createCategory",
        "summary": "Create category",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "image": {
                    "type": "string",
                    "default": ""
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published"],
                    "default": "draft"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["draft", "published"]
                    },
                    "productCount": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "image",
                    "status",
                    "productCount",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": ["error", "details"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/wishlist": {
      "get": {
        "tags": ["Wishlist"],
        "operationId": "getWishlist",
        "summary": "Get current wishlist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Wishlist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "productId": {
                            "type": "string"
                          },
                          "productName": {
                            "type": "string"
                          },
                          "productSlug": {
                            "type": "string"
                          },
                          "productImage": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "inStock": {
                            "type": "boolean"
                          },
                          "addedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "productId",
                          "productName",
                          "productSlug",
                          "productImage",
                          "price",
                          "inStock",
                          "addedAt"
                        ]
                      }
                    },
                    "itemCount": {
                      "type": "integer"
                    }
                  },
                  "required": ["id", "items", "itemCount"]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Wishlist"],
        "operationId": "clearWishlist",
        "summary": "Clear wishlist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Empty wishlist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "productId": {
                            "type": "string"
                          },
                          "productName": {
                            "type": "string"
                          },
                          "productSlug": {
                            "type": "string"
                          },
                          "productImage": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "inStock": {
                            "type": "boolean"
                          },
                          "addedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "productId",
                          "productName",
                          "productSlug",
                          "productImage",
                          "price",
                          "inStock",
                          "addedAt"
                        ]
                      }
                    },
                    "itemCount": {
                      "type": "integer"
                    }
                  },
                  "required": ["id", "items", "itemCount"]
                }
              }
            }
          }
        }
      }
    },
...
  }
}